Extract common behavior to a base
The `client_certificate`'s were sharing some common behavior,
this commit extract those behavior to a base class, and all
of the `client_certificate` will inherit from this one, so
we can easily put the common behavior here.
