+++ lib/digicert/client_certificate/email_security_plus.rb
@@ -1,20 +1,12 @@
-require "digicert/base_order"
+require "digicert/client_certificate/base"
 
 module Digicert
   module ClientCertificate
-    class EmailSecurityPlus < Digicert::BaseOrder
+    class EmailSecurityPlus < Digicert::ClientCertificate::Base
       private
 
       def certificate_type
         "client_email_security_plus"
-      end
-
-      def validate_certificate(common_name:, signature_hash:, emails:, **attrs)
-        attrs.merge(
-          emails: emails,
-          common_name: common_name,
-          signature_hash: signature_hash,
-        )
       end
     end
   end
