+++ lib/digicert/client_certificate/premium.rb
@@ -1,16 +1,16 @@
-require "digicert/base_order"
+require "digicert/client_certificate/base"
 
 module Digicert
   module ClientCertificate
-    class Premium < Digicert::BaseOrder
+    class Premium < Digicert::ClientCertificate::Base
       private
 
       def certificate_type
         "client_premium_sha2"
       end
 
-      def validate_certificate(emails:, **attributes)
-        super(attributes.merge(emails: emails))
+      def validate_certificate(csr:, **attributes)
+        super(attributes.merge(csr: csr))
       end
     end
   end
